requirements.txt
faster-whisper>=1.0.3
openwakeword>=0.6.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
pyyaml>=6.0.2
requests>=2.31.0
piper-tts>=1.2.0
onnxruntime>=1.17.0
